+++ readtog.php
@@ -1,60 +1,40 @@
 <?php
-// Let's crank up, conditionally, the lights at the north-end of the living room. 2 downlights, the end-table, and 3
-// bulbs in the tall dragonfly. 
+# Let's crank up, conditionally, the lights at the north-end of the living room. 2 downlights, the end-table, and 3
+# bulbs in the tall dragonfly. 
 include 'functions.php';
-$dl1 = 22; //ambiance only
-$dl2 = 23; //ambiance only
-$et1 = 12; //end-table - Hue
-$df1 = 13; //tall dragonfly lamp 1 - Hue
-$df2 = 14; 
-$df3 = 15;
+$dl1 = 22; #ambiance only
+$dl2 = 23; #ambiance only
+$huelights = array(12,13,14,15); # HS-capable lights. Let's save HueState for these, just in case they're in SP2 mode (or some other colour) already.
 
-// First, the downlights...
-if (restoreState($dl1)) {
-    // Restore done.
-}
-else {
+if(!checkState($dl1)) { # Key on the first DL. If it's not in-progress, assume none of the others are. (this is step 1)
     saveCTState($dl1);
-    oneOnSpotCool($dl1);
-}
-if (restoreState($dl2)) {
-    // Restore done.
-}
-else {
     saveCTState($dl2);
+    foreach($huelights as $id) {
+        saveHueState($id);
+    }
+    
+    # Now that they're all saved, let's shift 'em... 
+    oneOnSpotWarm($dl1);
+    oneOnSpotWarm($dl2);
+    foreach($huelights as $id) {
+        oneOnSpotWarm($id);
+    }
+}
+elseif(getCTState($dl1) === $ctWarm) { # we're already in read-mode with warm downlights (from step 1 above)
+    oneOnSpotCool($dl1);
     oneOnSpotCool($dl2);
+    foreach($huelights as $id) {
+        oneOnSpotCool($id);
+    }
+
+} 
+else { # we'll only get here if the downlights are $ctCool (not warm), so let's restore the previous state
+    restoreState($dl1);
+    restoreState($dl2);
+    foreach($huelights as $id) {
+        restoreState($id);
+    }
 }
 
-// Next, the end-table light. Probably not critical (I'm usually at the east end, but whatever, man)
-if (restoreState($et1)) {
-    // Restore done.
-}
-else {
-    saveHueState($et1);
-    oneOnBright($et1);
-}
-
-// Finally, the 3 bulbs in the tall dragonfly floor lamp.
-if (restoreState($df1)) {
-    // Restore done.
-}
-else {
-    saveHueState($df1);
-    oneOnBright($df1);
-}
-if (restoreState($df2)) {
-    // Restore done.
-}
-else {
-    saveHueState($df2);
-    oneOnBright($df2);
-}
-if (restoreState($df3)) {
-    // Restore done.
-}
-else {
-    saveHueState($df3);
-    oneOnBright($df3);
-}
 if(strpos($_SERVER['HTTP_USER_AGENT'], 'WebKit')) header('Location: ' . $_SERVER['readtog.php']);
 ?>